+++ apps/common-app/src/examples/RuntimeTests/ReJest/TestRunner/TestRunner.ts
@@ -18,7 +18,7 @@ import {
   EMPTY_LOG_PLACEHOLDER,
   applyMarkdown,
   color,
-  formatString,
+  formatTestName,
   indentNestingLevel,
 } from '../utils/stringFormatUtils';
 import type {
@@ -168,10 +168,10 @@ export class TestRunner {
           await testCase(example, index);
         };
         this.test(
-          formatString(name, example, index),
+          formatTestName(name, example, index),
           currentTestCase,
           decorator,
-          formatString(expectedWarning, example, index),
+          formatTestName(expectedWarning, example, index),
         );
       });
     };
@@ -183,7 +183,7 @@ export class TestRunner {
         const currentTestCase = async () => {
           await testCase(example, index);
         };
-        this.test(formatString(name, example, index), currentTestCase, decorator);
+        this.test(formatTestName(name, example, index), currentTestCase, decorator);
       });
     };
   }
